(Perf improvement) Keep dynamically allocated memory in filterPeaks().
Allocating and freeing memory added an extra ~10% time in the example
model included for testing.

So, now it allocates on first use or if it needs a larger buffers.
It never deallocates, but the memory consumption penalty is worth
the runtime performance improvement, as we expect the model to be
constantly running in the bacground.